src/main/java/nl/han/oose/dea/Entity/PlaylistEntity.java
package nl.han.oose.dea.Entity;

import java.util.List;

public class PlaylistEntity {
    private int id;
    private String name;
    private boolean owner;
    private List<TrackEntity> tracks;
    private int length;

    public PlaylistEntity() {
    }

    public PlaylistEntity(int id, String name, boolean owner) {
        this.id = id;
        this.name = name;
        this.owner = owner;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public boolean isOwner() {
        return owner;
    }

    public void setOwner(boolean owner) {
        this.owner = owner;
    }

    public List<TrackEntity> getTracks() {
        return tracks;
    }

    public void setTracks(List<TrackEntity> tracks) {
        this.tracks = tracks;
    }

    public int getLength() {
        return length;
    }

    public void setLength(int length) {
        this.length = length;
    }
}
